+++ php/src/Application/Command/SemanticLint/UnusedUseStatementAnalyzer.php
@@ -2,6 +2,7 @@
 
 namespace PhpIntegrator\Application\Command\SemanticLint;
 
+use PhpIntegrator\DocParser;
 use PhpIntegrator\TypeAnalyzer;
 
 /**
@@ -28,12 +29,13 @@ class UnusedUseStatementAnalyzer implements AnalyzerInterface
      * Constructor.
      *
      * @param TypeAnalyzer $typeAnalyzer
+     * @param DocParser    $docParser
      */
-    public function __construct(TypeAnalyzer $typeAnalyzer)
+    public function __construct(TypeAnalyzer $typeAnalyzer, DocParser $docParser)
     {
         $this->classUsageFetchingVisitor = new Visitor\ClassUsageFetchingVisitor();
         $this->useStatementFetchingVisitor = new Visitor\UseStatementFetchingVisitor();
-        $this->docblockClassUsageFetchingVisitor = new Visitor\DocblockClassUsageFetchingVisitor($typeAnalyzer);
+        $this->docblockClassUsageFetchingVisitor = new Visitor\DocblockClassUsageFetchingVisitor($typeAnalyzer, $docParser);
     }
 
     /**